feat(observability): M1 #12 run traces + structured logging
Migration 0005 adds the traces table (spine-only deps). project_run_trace
runs at settle in the worker: writes one trace per run (trace_kind, status,
started/ended, tags with provider/model + estimated input/output tokens +
cost) and advances the session token/cost rollups. Structured JSON logging
with tenant/run/session/request correlation contextvars and secret redaction
(config §3.5), configured via app lifespan + worker on_startup. Fixes
sessions.cost_usd_rollup mapping to Decimal (Numeric). generations and
audit_receipts are deferred to M2 (they FK to extractions/approval_envelopes
and audit_receipts needs the append-only trigger). Trace + formatter tests.

Copilot-Session: 44048257-622a-4ee7-be67-57691e3c8f8b

+++ backend/app/models/core.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 import datetime
 import uuid
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import (
@@ -117,7 +118,7 @@ class Session(Base):
     fence_token: Mapped[int] = mapped_column(BigInteger, server_default="0")
     input_tokens_rollup: Mapped[int] = mapped_column(BigInteger, server_default="0")
     output_tokens_rollup: Mapped[int] = mapped_column(BigInteger, server_default="0")
-    cost_usd_rollup: Mapped[float] = mapped_column(Numeric(20, 8), server_default="0")
+    cost_usd_rollup: Mapped[Decimal] = mapped_column(Numeric(20, 8), server_default="0")
     last_activity_at: Mapped[datetime.datetime | None] = mapped_column(DateTime(timezone=True))
     created_at = _ts()
     updated_at = _ts()
